+++ crates/c-api/resvg.h
@@ -431,7 +431,7 @@ bool resvg_get_node_transform(const resvg_render_tree *tree,
                               resvg_transform *transform);
 
 /**
- * @brief Returns node's bounding box by ID.
+ * @brief Returns node's bounding box in canvas coordinates by ID.
  *
  * @param tree Render tree.
  * @param id Node's ID. Must not be NULL.
@@ -441,6 +441,18 @@ bool resvg_get_node_transform(const resvg_render_tree *tree,
  * @return `false` if ID is an empty string
  */
 bool resvg_get_node_bbox(const resvg_render_tree *tree, const char *id, resvg_rect *bbox);
+
+/**
+ * @brief Returns node's bounding box, including stroke, in canvas coordinates by ID.
+ *
+ * @param tree Render tree.
+ * @param id Node's ID. Must not be NULL.
+ * @param bbox Node's bounding box.
+ * @return `false` if a node with such an ID does not exist
+ * @return `false` if ID isn't a UTF-8 string.
+ * @return `false` if ID is an empty string
+ */
+bool resvg_get_node_stroke_bbox(const resvg_render_tree *tree, const char *id, resvg_rect *bbox);
 
 /**
  * @brief Destroys the #resvg_render_tree.